requirements.txt
streamlit
pandas
gspread
google-auth
fpdf
plotly
requests